+++ examples/resources/pass_password/import.sh
@@ -1,0 +1,1 @@
+terraform import pass_password.example Utvikling/Azure/db_password